% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dead_air_rows.R
\name{add_dead_air_rows}
\alias{add_dead_air_rows}
\title{Add Dead Air Rows}
\usage{
add_dead_air_rows(df, dead_air_name = "dead_air")
}
\arguments{
\item{df}{A tibble containing the comments from a Zoom recording transcript.}

\item{dead_air_name}{Character string to label the \code{name} column in the added rows. Defaults to 'dead_air'.}
}
\value{
A tibble containing the comments from a Zoom recording transcript, with rows added for dead air.
}
\description{
Take a tibble containing the comments from a Zoom recording transcript and return a tibble that adds rows for any time between transcribed comments, labeled with the \code{dead_air_name} provided (or the default value of 'dead_air').  The resulting tibble will have rows accounting for the time from the beginning of the first comment to the end of the last one.
}
\examples{
add_dead_air_rows(df = "NULL")

}
