% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_transcript.R
\name{consolidate_transcript}
\alias{consolidate_transcript}
\title{Consolidate Transcript}
\usage{
consolidate_transcript(df, max_pause_sec = 1)
}
\arguments{
\item{df}{A tibble containing the comments from a Zoom recording transcript.}

\item{max_pause_sec}{Maximum pause between comments to be consolidated.  If
the raw comments from the Zoom recording transcript contain 2 consecutive
comments from the same speaker, and the time between the end of the first
comment and start of the second comment is less than `max_pause_sec`
seconds, then the comments will be consolidated.  If the time between the
comments is larger, they will not be consolidated. Defaults to 1.}
}
\value{
A tibble containing consolidated comments from a Zoom recording
  transcript.
}
\description{
Take a tibble containing the comments from a Zoom recording transcript and return a tibble that consolidates all consecutive comments from the same speaker where the time between the end of the first comment and start of the second comment is less than `max_pause_sec` seconds.  This function addresses an issue with the Zoom transcript where the speaker is speaking a continuous sentence, but the Zoom transcript will cut the comment into two lines.
For example, a comment of "This should be a single sentence." is often split into "This should be" and "a single sentence".  This function stitches those together into "This should be a single sentence." where the `start` time of the consolidated comment will be the beginning of the first row and the `end` time of the consolidated comment will be the ending of the last row.
}
\examples{
consolidate_transcript(df = "NULL")

}
