% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_diagnostics.R
\name{diag_cat}
\alias{diag_cat}
\title{Conditionally emit cat-style output when verbose is enabled or in interactive sessions}
\usage{
diag_cat(...)
}
\description{
Conditionally emit cat-style output when verbose is enabled or in interactive sessions
}
\keyword{internal}
