% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_diagnostics.R
\name{diag_cat_if}
\alias{diag_cat_if}
\title{Conditionally emit cat output if a local verbose flag is TRUE or global verbose is enabled}
\usage{
diag_cat_if(verbose_flag, ...)
}
\arguments{
\item{verbose_flag}{Logical flag controlling local verbosity}
}
\description{
Conditionally emit cat output if a local verbose flag is TRUE or global verbose is enabled
}
\keyword{internal}
