% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_diagnostics.R
\name{diag_message}
\alias{diag_message}
\title{Conditionally emit a message when verbose is enabled}
\usage{
diag_message(...)
}
\description{
Conditionally emit a message when verbose is enabled
}
\keyword{internal}
