% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_privacy_compliance.R
\name{extract_character_values}
\alias{extract_character_values}
\title{Extract Character Values from Data Object}
\usage{
extract_character_values(data)
}
\arguments{
\item{data}{Data object to extract character values from}
}
\value{
Character vector of all character values found
}
\description{
Internal function to extract all character values from various data types
for privacy validation.
}
\keyword{internal}
