% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fliwc.R
\name{fliwc}
\alias{fliwc}
\title{Faculty Linguistic Inquiry and Word Count}
\usage{
fliwc(transcript_file_path, names_exclude = c("dead_air"))
}
\arguments{
\item{transcript_file_path}{File path of a .vtt file of a Zoom recording
transcript.}

\item{names_exclude}{Character vector of names to exclude from the results.
Defaults to 'c("dead_air")'}
}
\value{
A tibble containing summary metrics by speaker from a Zoom recording
  transcript
}
\description{
Process a Zoom recording transcript and return summary metrics by speaker
}
\details{
Original code posted by Conor Healy:
https://ucbischool.slack.com/archives/C02A36407K9/p1631855705002000 Addition
of `wordcount`, `wordcount_perc`, and `wpm` by Brooks Ambrose:
https://gist.github.com/brooksambrose/1a8a673eb3bf884c1868ad4d80f08246
}
\examples{
fliwc(transcript_file_path = 'NULL')
}
