% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_transcript_files.R
\name{summarize_transcript_files}
\alias{summarize_transcript_files}
\title{Summarize Transcript Files}
\usage{
summarize_transcript_files(
  df_transcript_list,
  data_folder = "data",
  transcripts_folder = "transcripts",
  names_to_exclude = NULL
)
}
\arguments{
\item{df_transcript_list}{A data.frame listing the transcript files from the
zoom recordings loaded from the cloud recording csvs and transcripts.}

\item{data_folder}{Overall data folder for your recordings and data. Defaults
to 'data'}

\item{transcripts_folder}{specific subfolder of the data folder where you
will store the cloud recording csvs and transcripts}

\item{names_to_exclude}{Character vector of names to exclude from the results.
Defaults to NULL}
}
\value{
A tibble containing session details and summary metrics by speaker
  for all class sessions in the tibble provided.
}
\description{
Load Zoom Transcript file and Run Faculty Linguistic Inquiry and Word Count
on those sessions
}
\examples{
summarize_transcript_files(df_transcript_list = NULL)
}
