% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_process_zoom_transcript.R
\name{load_and_process_zoom_transcript}
\alias{load_and_process_zoom_transcript}
\title{Load and Process Zoom Transcript}
\usage{
load_and_process_zoom_transcript(
  transcript_file_path,
  consolidate_comments = TRUE,
  max_pause_sec = 1,
  add_dead_air = TRUE,
  dead_air_name = "dead_air",
  na_name = "unknown"
)
}
\arguments{
\item{transcript_file_path}{File path of a .vtt file of a Zoom recording
transcript.}

\item{consolidate_comments}{Set to `TRUE` to consolidate consecutive comments
from the same speaker with gaps of less than `max_pause_sec`. `FALSE`
returns the results from the raw transcript.  Defaults to `TRUE`}

\item{max_pause_sec}{Maximum pause between comments to be consolidated.  If
the raw comments from the Zoom recording transcript contain 2 consecutive
comments from the same speaker, and the time between the end of the first
comment and start of the second comment is less than `max_pause_sec`
seconds, then the comments will be consolidated.  If the time between the
comments is larger, they will not be consolidated. Defaults to 1.}

\item{add_dead_air}{Set to `TRUE` to adds rows for any time between
transcribed comments, labeled with the `dead_air_name` provided (or the
default value of 'dead_air').  The resuting tibble will have rows
accounting for the time from the beginning of the first comment to the end
of the last one. Defaults to `TRUE`.}

\item{dead_air_name}{Character string to label the `name` column in any rows
added for dead air. Defaults to 'dead_air'.}

\item{na_name}{Character string to label the `name` column in any rows where
the transcript `name` is `NA`. Defaults to 'unknown'.}
}
\value{
A tibble containing the comments from a Zoom recording transcript
}
\description{
DEPRICATED - Use process_zoom_transcript() instead. - Load a Zoom recording transcript and return tibble containing the comments from a Zoom recording transcript
}
\details{
Original code posted by Conor Healy:
https://ucbischool.slack.com/archives/C02A36407K9/p1631855705002000 Addition
of `wordcount`, `wordcount_perc`, and `wpm` by Brooks Ambrose:
https://gist.github.com/brooksambrose/1a8a673eb3bf884c1868ad4d80f08246
}
\examples{
# Load a sample transcript from the package's extdata directory
transcript_file <- system.file("extdata/transcripts/GMT20240124-202901_Recording.transcript.vtt",
  package = "zoomstudentengagement"
)
load_and_process_zoom_transcript(transcript_file_path = transcript_file)

}
