% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cancelled_classes.R
\name{load_cancelled_classes}
\alias{load_cancelled_classes}
\title{Load Cancelled Classes csv file}
\usage{
load_cancelled_classes(
  data_folder = ".",
  cancelled_classes_file = "cancelled_classes.csv",
  cancelled_classes_col_types = "ccccccccnnnncTTcTTccci",
  write_blank_cancelled_classes = FALSE
)
}
\arguments{
\item{data_folder}{Overall data folder for your recordings and data. Defaults
to 'data'}

\item{cancelled_classes_file}{File name of the csv file of cancelled classes.
Defaults to 'cancelled_classes.csv'}

\item{cancelled_classes_col_types}{column types in the csv file of cancelled
classes. Defaults to 'ccccccccnnnncTTcTTccci'}

\item{write_blank_cancelled_classes}{Logical. If TRUE and the file doesn't exist,
creates a blank cancelled classes file. Defaults to FALSE}
}
\value{
A tibble listing the cancelled class sessions for scheduled classes
where a zoom recording is not expected.
}
\description{
This function creates a tibble from a provided csv file of cancelled class
sessions for scheduled classes where a zoom recording is not expected.
}
\examples{
load_cancelled_classes()
}
