% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_section_names_lookup.R
\name{load_section_names_lookup}
\alias{load_section_names_lookup}
\title{Load Section Names Lookup File}
\usage{
load_section_names_lookup(
  data_folder = "data",
  names_lookup_file = "section_names_lookup.csv",
  section_names_lookup_col_types = "cccdcccd"
)
}
\arguments{
\item{data_folder}{overall data folder for your recordings}

\item{names_lookup_file}{File name of the csv file of customized student names by section
Defaults to 'section_names_lookup.csv'}

\item{section_names_lookup_col_types}{column types in the csv file of customized student names by section. Defaults to 'cccdcccd'}
}
\value{
A tibble of customized student names by section.
}
\description{
This function creates a tibble from a provided file of customized student names by section.
If the file does not exist, the function prints an error and creates an empty tibble using `make_blank_section_names_lookup_csv()`.
}
\examples{
load_section_names_lookup()

}
