% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_transcript_files_list.R
\name{load_transcript_files_list}
\alias{load_transcript_files_list}
\title{Load Zoom Recording Transcript Files List}
\usage{
load_transcript_files_list(
  data_folder = "data",
  transcripts_folder = "transcripts",
  transcript_files_names_pattern = "GMT\\\\d{8}-\\\\d{6}_Recording",
  dt_extract_pattern = "(?<=GMT)\\\\d{8}",
  transcript_file_extension_pattern = ".transcript",
  closed_caption_file_extension_pattern = ".cc",
  recording_start_pattern = "(?<=GMT)\\\\d{8}-\\\\d{6}",
  recording_start_format = "\%Y\%m\%d-\%H\%M\%S",
  start_time_local_tzone = "America/Los_Angeles"
)
}
\arguments{
\item{data_folder}{Overall data folder for your recordings and data. Defaults
to 'data'}

\item{transcripts_folder}{specific subfolder of the data folder where you
will store the cloud recording csvs and transcripts}

\item{transcript_files_names_pattern}{REGEX pattern used to match the
transcript file names. Defaults to 'GMT\\d{8}-\\d{6}_Recording'}

\item{dt_extract_pattern}{REGEX pattern used to extract the date of the
transcript from the transcript file name. Defaults to '(?<=GMT)\\d{8}'}

\item{transcript_file_extension_pattern}{REGEX pattern used to identify
transcript files (as opposed to chat or closed caption files). Defaults to
'.transcript'}

\item{closed_caption_file_extension_pattern}{REGEX pattern used to identify
closed caption files (as opposed to chat or transcript files). Defaults to
'.cc'}

\item{recording_start_pattern}{REGEX pattern used to extract the recording
start time of the transcript from the transcript file name. Defaults to
'(?<=GMT)\\d{8}-\\d{6}'}

\item{recording_start_format}{Pattern used to parse the format of the
recording start time of the transcript. Defaults to '\%Y\%m\%d-\%H\%M\%S'}

\item{start_time_local_tzone}{Local time zone of the recording start time of
the transcript. Defaults to 'America/Los_Angeles'}
}
\value{
A data.frame listing the transcript files from the zoom recordings
  loaded from the cloud recording csvs and transcripts.
}
\description{
This function creates a data.table from a provided folder including
transcript files of Zoom recordings.
}
\details{
## Download Transcripts 1. Go to
[https://www.zoom.us/recording](https://www.zoom.us/recording) 2. Click on
each individual record to go to the page for that recording 3. Download the
Audio Transcript and Chat File for each + Chat:
'GMT\\d{8}-\\d{6}_Recording.cc.vtt' + Transcript:
'GMT\\d{8}-\\d{6}_Recording.transcript.vtt' 4. Copy the Audio Transcript and
Chat Files to 'data/transcripts/' (or whatever path you identify in the
`data_folder` and `transcripts_folder` parameters).
}
\examples{
load_transcript_files_list()
}
