% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferpa_compliance.R
\name{log_ferpa_compliance_check}
\alias{log_ferpa_compliance_check}
\title{Log FERPA Compliance Check}
\usage{
log_ferpa_compliance_check(
  compliant,
  pii_detected,
  institution_type,
  timestamp = Sys.time()
)
}
\arguments{
\item{compliant}{Whether the data is FERPA compliant}

\item{pii_detected}{Number of PII fields detected}

\item{institution_type}{Type of institution}

\item{timestamp}{When the check was performed}
}
\description{
Internal function to log FERPA compliance checks for audit and institutional
review purposes.
}
\keyword{internal}
