% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blank_cancelled_classes_df.R
\name{make_blank_cancelled_classes_df}
\alias{make_blank_cancelled_classes_df}
\title{Make Cancelled Classes Tibble
This function creates an empty tibble for recording of cancelled class
sessions for scheduled classes where a Zoom recording is not expected.}
\usage{
make_blank_cancelled_classes_df()
}
\value{
An empty tibble for recording of cancelled class sessions for scheduled classes
  where a zoom recording is not expected.
}
\description{
Make Cancelled Classes Tibble
This function creates an empty tibble for recording of cancelled class
sessions for scheduled classes where a Zoom recording is not expected.
}
\examples{
make_blank_cancelled_classes_df()
}
