% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metrics_lookup_df.R
\name{make_metrics_lookup_df}
\alias{make_metrics_lookup_df}
\title{Make Metrics Lookup DF}
\usage{
make_metrics_lookup_df()
}
\value{
A tibble including metric labels and metric descriptions
}
\description{
This function creates a tibble for the metrics that will be analyzed in the
analysis, with text descriptions to go along with the metric labels
}
\examples{
make_metrics_lookup_df()
}
\keyword{metrics}
