% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_analysis_template.R
\name{make_new_analysis_template}
\alias{make_new_analysis_template}
\title{Make New Analysis Template}
\usage{
make_new_analysis_template(
  new_template_file_name = "new_analysis.Rmd",
  template_file = system.file("new_analysis_template.Rmd", package =
    "zoomstudentengagement")
)
}
\arguments{
\item{new_template_file_name}{The file name of the created report template.}

\item{template_file}{The file name of the template file. This defaults to the
`new_analysis_template.Rmd` file in the `zoomstudentengagement` package.}
}
\value{
TRUE if the file was copied, FALSE otherwise.
}
\description{
This function copies a template R Markdown file (the `new_analysis_template.Rmd` file
from the `zoomstudentengagement` package) and saves it as a new .Rmd file that can be
used as a starting point for running analyses with the `zoomstudentengagement` library.
}
\examples{
\dontrun{
make_new_analysis_template(
  new_template_file_name = "my_analysis.Rmd"
)
}
}
