% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_roster_small.R
\name{make_roster_small}
\alias{make_roster_small}
\title{Make a Smaller DF of the Student Roster}
\usage{
make_roster_small(roster_df)
}
\arguments{
\item{roster_df}{A tibble listing the students enrolled in the class or
classes with a small subset of the roster columns.}
}
\value{
A tibble listing the students enrolled in the class or classes with a
  small subset of the roster columns.
}
\description{
This function creates a tibble that includes rows for each students enrolled
in the class or classes, with a small subset of the roster columns.
}
\examples{
make_roster_small(roster_df = load_roster())
}
\keyword{roster}
