% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sections_df.R
\name{make_sections_df}
\alias{make_sections_df}
\title{Make a DF of Class Sections from the Student Roster}
\usage{
make_sections_df(roster_df)
}
\arguments{
\item{roster_df}{A tibble listing the students enrolled in the class or classes.}
}
\value{
A tibble including metric labels and metric descriptions
}
\description{
This function creates a tibble that includes rows for each section (grouped by dept and course number) and student count in each
}
\examples{
make_sections_df(roster_df = load_roster())
}
\keyword{sections}
