% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_semester_df.R
\name{make_semester_df}
\alias{make_semester_df}
\title{Make Semester DF}
\usage{
make_semester_df(
  semester_units = 14,
  class_start_time_gmt = "04:00:00",
  class_duration_min = 90
)
}
\arguments{
\item{semester_units}{Number of units/weeks/classes in the semester. Defaults
to 14.}

\item{class_start_time_gmt}{Start time of the class in GMT, formatted as
'HH:MM:SS'. Defaults to '04:00:00'.}

\item{class_duration_min}{Number of units/weeks/classes in the semester.
Defaults to 14.}
}
\value{
A tibble including number of units, with class start time, end time,
  and duration.
}
\description{
This function creates a tibble for the units in the semester that correspond
to expected number of units, with class start time, end time, and duration.
}
\examples{
make_semester_df()
}
\keyword{semester}
