% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_student_roster_sessions.R
\name{make_student_roster_sessions}
\alias{make_student_roster_sessions}
\title{Make a DF of the Student Roster With Rows for Each Recorded Class Section}
\usage{
make_student_roster_sessions(transcripts_list_df, roster_small_df)
}
\arguments{
\item{transcripts_list_df}{A tibble listing the class sessions with corresponding transcript
files or placeholders for cancelled classes.}

\item{roster_small_df}{A tibble listing the students enrolled in the class or classes with a
small subset of the roster columns.}
}
\value{
A tibble listing the students enrolled in the class or classes, with rows for each recorded class section for each student.
}
\description{
This function creates a tibble from a provided tibble students enrolled
in the class or classes (`roster_small_df`) and a tibble of class sessions with corresponding transcript
  files or placeholders for cancelled classes (`transcripts_list_df`).
}
\examples{
make_student_roster_sessions(
  transcripts_list_df = join_transcripts_list(
    df_zoom_recorded_sessions = load_zoom_recorded_sessions_list(),
    df_transcript_files = load_transcript_files_list(),
    df_cancelled_classes = load_cancelled_classes()
  ),
  roster_small_df = make_roster_small(
    roster_df = load_roster()
  )
)
}
