% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_template_rmd.R
\name{make_template_rmd}
\alias{make_template_rmd}
\title{Write Section Names Lookup}
\usage{
make_template_rmd(
  new_template_file_name = "README.Rmd",
  template_file = paste0(system.file("", package = "zoomstudentengagement"),
    "README.Rmd")
)
}
\arguments{
\item{new_template_file_name}{The file name of the created report template.}

\item{template_file}{The file name of the template file. This defaults to the
`README.Rmd` file in the `zoomstudentengagement` documentation.}
}
\value{
TRUE if the file was copied, FALSE otherwise.
}
\description{
This function a template file (generally the `README.Rmd` file in the
`zoomstudentengagement` documentation) and saves it as a .Rmd file that can be
used to run analyses with the `zoomstudentengagement` library.
}
\examples{
make_template_rmd(
  new_template_file_name = "README.Rmd",
  template_file = "README.Rmd"
)
}
