% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_user_names_by_metric.R
\name{mask_user_names_by_metric}
\alias{mask_user_names_by_metric}
\title{Mask User Names by Metric}
\usage{
mask_user_names_by_metric(df, metric = "session_ct", target_student = "")
}
\arguments{
\item{df}{a tibble that summarizes results at the level of the class section
and student.  This tibble will have the student names replaced by the
ranking of the student.  If a `target_student` preferred name is provided,
that student's name will be bolded using markdown syntax and not masked.}

\item{metric}{Label of the metric to use to order the students. Defaults to
'session_ct'.}

\item{target_student}{preferred student name of an individual student that
will be bolded using markdown syntax and not masked. Defaults to ''.}
}
\value{
a tibble that summarizes results at the level of the class section
  and student, with student names masked by the ranking of the student.
}
\description{
Mask User Names by Metric
}
\examples{
mask_user_names_by_metric(
  make_transcripts_summary_df(
    make_transcripts_session_summary_df(
      clean_names_df = make_clean_names_df(
        data_folder = "data",
        section_names_lookup_file = "section_names_lookup.csv",
        transcripts_fliwc_df = fliwc_transcript_files(df_transcript_list = NULL),
        roster_sessions = make_student_roster_sessions(
          transcripts_list_df = join_transcripts_list(
            df_zoom_recorded_sessions = load_zoom_recorded_sessions_list(),
            df_transcript_files = load_transcript_files_list(),
            df_cancelled_classes = load_cancelled_classes()
          ),
          roster_small_df = make_roster_small(
            roster_df = load_roster()
          )
        )
      )
    )
  )
)

}
