% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_section_names_lookup.R
\name{validate_lookup_file_format}
\alias{validate_lookup_file_format}
\title{Validate Lookup File Format}
\usage{
validate_lookup_file_format(lookup_data)
}
\arguments{
\item{lookup_data}{Data frame containing lookup information}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Internal function to validate the format of section names lookup files.
}
\keyword{internal}
