% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_section_names_lookup.R
\name{write_section_names_lookup}
\alias{write_section_names_lookup}
\title{Write Section Names Lookup}
\usage{
write_section_names_lookup(
  clean_names_df,
  data_folder = "data",
  section_names_lookup_file = "section_names_lookup.csv"
)
}
\arguments{
\item{clean_names_df}{A tibble containing session details and summary metrics
by speaker for all class sessions (and placeholders for missing sections),
including customized student names.}

\item{data_folder}{Overall data folder for your recordings and data. Defaults
to 'data'}

\item{section_names_lookup_file}{File name of the csv file of customized
student names by section Defaults to 'section_names_lookup.csv'}
}
\value{
A tibble coresponding to the csv file saved, which is a sorted subset
  of the provided tibble containing session details and summary metrics by
  speaker for all class sessions (and placeholders for missing sections),
  including customized student names.
}
\description{
This function takes a tibble containing session details and summary metrics
by speaker for all class sessions (and placeholders for missing sections),
including customized student names (`clean_names_df`) and saves a subset as a
csv file with the specified file name (`section_names_lookup_file`) to the
specified data folder (`data_folder`).
}
\examples{
write_section_names_lookup(
  clean_names_df = make_clean_names_df(
    data_folder = "data",
    section_names_lookup_file = "section_names_lookup.csv",
    transcripts_metrics_df = summarize_transcript_files(df_transcript_list = NULL),
    roster_sessions = make_student_roster_sessions(
      transcripts_list_df = join_transcripts_list(
        df_zoom_recorded_sessions = load_zoom_recorded_sessions_list(),
        df_transcript_files = load_transcript_files_list(),
        df_cancelled_classes = load_cancelled_classes()
      ),
      roster_small_df = make_roster_small(
        roster_df = load_roster()
      )
    )
  ),
  data_folder = "data",
  section_names_lookup_file = "section_names_lookup.csv"
)
}
