% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomstudentengagement-package.R
\docType{package}
\name{zoomstudentengagement-package}
\alias{zoomstudentengagement}
\alias{zoomstudentengagement-package}
\title{zoomstudentengagement: Zoom Student Engagement}
\description{
Takes Zoom transcripts and student rosters, converts them for analysis, and analyzes the results for student engagement, with a particular focus on participation equity.
}
\keyword{internal}
