% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zoom_transcript.R
\name{load_zoom_transcript}
\alias{load_zoom_transcript}
\title{Faculty Linguistic Inquiry and Word Count}
\usage{
load_zoom_transcript(transcript_file_path)
}
\arguments{
\item{transcript_file_path}{File path of a .vtt file of a Zoom recording
transcript.}
}
\value{
A tibble containing the comments from a Zoom recording
  transcript
}
\description{
Load a Zoom recording transcript and return tibble containing the comments from a Zoom recording transcript
}
\details{
Original code posted by Conor Healy:
https://ucbischool.slack.com/archives/C02A36407K9/p1631855705002000 Addition
of `wordcount` by Brooks Ambrose:
https://gist.github.com/brooksambrose/1a8a673eb3bf884c1868ad4d80f08246
}
\examples{
load_zoom_transcript(transcript_file_path = "NULL")

}
