% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blank_section_names_lookup_csv.R
\name{make_blank_section_names_lookup_csv}
\alias{make_blank_section_names_lookup_csv}
\title{Make Blank Section Names Lookup Tibble}
\usage{
make_blank_section_names_lookup_csv()
}
\value{
An empty tibble for customization of student names by section.
}
\description{
This function creates an empty tibble for customization of student names by section.
}
\examples{
make_blank_section_names_lookup_csv()
}
