% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_names_df.R
\name{make_clean_names_df}
\alias{make_clean_names_df}
\title{Make Smaller DF of the Student Roster}
\usage{
make_clean_names_df(
  data_folder = "data",
  section_names_lookup_file = "section_names_lookup.csv",
  transcripts_fliwc_df,
  roster_sessions
)
}
\arguments{
\item{data_folder}{overall data folder for your recordings. Defaults to 'data'}

\item{section_names_lookup_file}{File name of the csv file of customized student names by section
Defaults to 'section_names_lookup.csv'}

\item{transcripts_fliwc_df}{A tibble containing session details and summary metrics by speaker
for all class sessions in the tibble provided.}

\item{roster_sessions}{A tibble listing the students enrolled in the class or classes, with rows for each recorded class section for each student.}
}
\value{
A tibble containing session details and summary metrics by speaker
  for all class sessions (and placeholders for missing sections), including customized student names.
}
\description{
This function creates a tibble from the joining of a tibble of customized student names by section (`section_names_lookup_file` in the `data_folder` folder),
a tibble containing session details and summary metrics by speaker
  for all class sessions (`transcripts_fliwc_df`), and A tibble listing the students enrolled in the class or classes, with rows for each recorded class section for each student
(`roster_sessions`) into a single tibble.
}
\examples{
make_clean_names_df(
  data_folder = "data",
  section_names_lookup_file = "section_names_lookup.csv",
  transcripts_fliwc_df = fliwc_transcript_files(df_transcript_list = NULL),
  roster_sessions = make_student_roster_sessions(
    transcripts_list_df = join_transcripts_list(
      df_zoom_recorded_sessions = load_zoom_recorded_sessions_list(),
      df_transcript_files = load_transcript_files_list(),
      df_cancelled_classes = load_cancelled_classes()
    ),
    roster_small_df = make_roster_small(
      roster_df = load_roster()
    )
  )
)

}
