% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_users_by_metric.R
\name{plot_users_by_metric}
\alias{plot_users_by_metric}
\title{Plot Users by Metric}
\usage{
plot_users_by_metric(
  transcripts_summary_df,
  metric = "session_ct",
  metrics_lookup_df = make_metrics_lookup_df()
)
}
\arguments{
\item{transcripts_summary_df}{a tibble that summarizes results at the level of the class section and preferred student name.}

\item{metric}{Label of the metric to plot. Defaults to 'session_ct'.}

\item{metrics_lookup_df}{A tibble including metric labels and metric descriptions.  Defaults to run `make_metrics_lookup_df()`}
}
\value{
A ggplot of the provided metrics by students from the provided tibble
}
\description{
Plot Users by Metric
}
\examples{
plot_users_by_metric(
  make_transcripts_summary_df(
    make_transcripts_session_summary_df(
      clean_names_df = make_clean_names_df(
        data_folder = 'data',
        section_names_lookup_file = 'section_names_lookup.csv',
        transcripts_fliwc_df = fliwc_transcript_files(df_transcript_list = NULL),
        roster_sessions = student_roster_sessions(
          transcripts_list_df = join_transcripts_list(
            df_zoom_recorded_sessions = load_zoom_recorded_sessions_list(),
            df_transcript_files = load_transcript_files_list(),
            df_cancelled_classes = load_cancelled_classes()
          ),
          roster_small_df = make_roster_small(
            roster_df = load_roster()
          )
        )
      )
    )
  )
)
}
