% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_users_masked_section_by_metric.R
\name{plot_users_masked_section_by_metric}
\alias{plot_users_masked_section_by_metric}
\title{Plot Users by Metric}
\usage{
plot_users_masked_section_by_metric(df, metric = "session_ct")
}
\arguments{
\item{df}{a tibble that summarizes results at the level of the class section and student.  This tibble will have the student names replaced by the ranking of the student.}

\item{metric}{Label of the metric to plot. Defaults to 'session_ct'.}
}
\value{
A ggplot of the provided metrics by students from the provided tibble, with student names masked by the ranking of the student.
}
\description{
Plot Users by Metric
}
\examples{
plot_users_masked_section_by_metric(
  make_transcripts_summary_df(
    make_transcripts_session_summary_df(
      clean_names_df = make_clean_names_df(
        data_folder = "data",
        section_names_lookup_file = "section_names_lookup.csv",
        transcripts_fliwc_df = fliwc_transcript_files(df_transcript_list = NULL),
        roster_sessions = student_roster_sessions(
          transcripts_list_df = join_transcripts_list(
            df_zoom_recorded_sessions = load_zoom_recorded_sessions_list(),
            df_transcript_files = load_transcript_files_list(),
            df_cancelled_classes = load_cancelled_classes()
          ),
          roster_small_df = make_roster_small(
            roster_df = load_roster()
          )
        )
      )
    )
  )
)

}
