% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_transcripts_summary.R
\name{write_transcripts_summary}
\alias{write_transcripts_summary}
\title{Write Transcripts Summary}
\usage{
write_transcripts_summary(
  transcripts_summary_df,
  data_folder = "data",
  transcripts_summary_file = "transcripts_summary.csv"
)
}
\arguments{
\item{transcripts_summary_df}{a tibble that summarizes results at the level of the class section and preferred student name.}

\item{data_folder}{Overall data folder for your recordings and data. Defaults
to 'data'}

\item{transcripts_summary_file}{File name of the csv file of summary
results at the level of the class section and preferred student name. Defaults to
'transcripts_summary.csv'}
}
\value{
A tibble coresponding to the csv file saved, which is a sorted subset
  of the provided tibble containing summary metrics by
  speaker for all class sections, including customized student names.
}
\description{
This function takes a  tibble that summarizes results at the level of the class section and preferred student name (`transcripts_summary_df`),
and saves it as a csv file with the specified file name
(`transcripts_summary_file`) to the specified data folder
(`data_folder`).
}
\examples{
write_transcripts_summary(
  make_transcripts_summary_df(
    make_transcripts_session_summary_df(
      clean_names_df = make_clean_names_df(
        data_folder = "data",
        section_names_lookup_file = "section_names_lookup.csv",
        transcripts_fliwc_df = fliwc_transcript_files(df_transcript_list = NULL),
        roster_sessions = student_roster_sessions(
          transcripts_list_df = join_transcripts_list(
            df_zoom_recorded_sessions = load_zoom_recorded_sessions_list(),
            df_transcript_files = load_transcript_files_list(),
            df_cancelled_classes = load_cancelled_classes()
          ),
          roster_small_df = make_roster_small(
            roster_df = load_roster()
          )
        )
      )
    )
  )
)
}
